requirements.txt
python-dotenv==1.0.0
uvicorn==0.21.1
fastapi==0.95.1
openai==0.27.4
tiktoken==0.3.3
langchain==0.0.144
faiss-cpu==1.7.3
